+++ src/PhoneNumber/ValidPhoneNumber.php
@@ -68,15 +68,15 @@ class ValidPhoneNumber implements PhoneNumberInterface, JsonSerializable, String
         $this->verifyPhoneNumberData($validatedPhoneNumberData);
 
         $this->valid               = (bool) $validatedPhoneNumberData->valid;
-        $this->number              = $validatedPhoneNumberData->number;
-        $this->localFormat         = $validatedPhoneNumberData->local_format;
-        $this->internationalFormat = $validatedPhoneNumberData->international_format;
-        $this->countryPrefix       = $validatedPhoneNumberData->country_prefix;
-        $this->countryCode         = $validatedPhoneNumberData->country_code;
-        $this->countryName         = $validatedPhoneNumberData->country_name;
-        $this->location            = $validatedPhoneNumberData->location;
-        $this->carrier             = $validatedPhoneNumberData->carrier;
-        $this->lineType            = $validatedPhoneNumberData->line_type;
+        $this->number              = (string) $validatedPhoneNumberData->number;
+        $this->localFormat         = (string) $validatedPhoneNumberData->local_format;
+        $this->internationalFormat = (string) $validatedPhoneNumberData->international_format;
+        $this->countryPrefix       = (string) $validatedPhoneNumberData->country_prefix;
+        $this->countryCode         = (string) $validatedPhoneNumberData->country_code;
+        $this->countryName         = (string) $validatedPhoneNumberData->country_name;
+        $this->location            = (string) $validatedPhoneNumberData->location;
+        $this->carrier             = (string) $validatedPhoneNumberData->carrier;
+        $this->lineType            = (string) $validatedPhoneNumberData->line_type;
     }
 
     /**